+++ livedata/src/main/java/io/getstream/chat/android/livedata/ChatNotificationConfigOffline.kt
@@ -1,11 +1,10 @@
 package io.getstream.chat.android.livedata
 
 import android.content.Context
-import android.content.Intent
 import androidx.core.content.ContextCompat
 import com.google.firebase.messaging.RemoteMessage
 import io.getstream.chat.android.client.notifications.options.ChatNotificationConfig
-import io.getstream.chat.android.livedata.service.SyncService
+import io.getstream.chat.android.livedata.service.sync.SyncService
 
 /**
  * Extend the ChatNotificationConfig to start the sync service when receiving new messages
